+++ pkg/fleet/installer/config/config.go
@@ -7,24 +7,16 @@
 package config
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"os"
 	"path"
 	"path/filepath"
-	"runtime"
 	"strings"
 
 	patch "gopkg.in/evanphx/json-patch.v4"
 	"gopkg.in/yaml.v3"
-
-	"github.com/DataDog/datadog-agent/pkg/fleet/installer/symlink"
-)
-
-const (
-	deploymentIDFile = ".deployment-id"
 )
 
 // FileOperationType is the type of operation to perform on the config.
@@ -55,89 +47,6 @@ type State struct {
 	ExperimentDeploymentID string
 }
 
-// GetState returns the state of the directories.
-func (d *Directories) GetState() (State, error) {
-	stablePath := filepath.Join(d.StablePath, deploymentIDFile)
-	experimentPath := filepath.Join(d.ExperimentPath, deploymentIDFile)
-	stableDeploymentID, err := os.ReadFile(stablePath)
-	if err != nil && !os.IsNotExist(err) {
-		return State{}, err
-	}
-	experimentDeploymentID, err := os.ReadFile(experimentPath)
-	if err != nil && !os.IsNotExist(err) {
-		return State{}, err
-	}
-	stableExists := len(stableDeploymentID) > 0
-	experimentExists := len(experimentDeploymentID) > 0
-	// If experiment is symlinked to stable, it means the experiment is not installed.
-	if stableExists && experimentExists && isSameFile(stablePath, experimentPath) {
-		experimentDeploymentID = nil
-	}
-	return State{
-		StableDeploymentID:     string(stableDeploymentID),
-		ExperimentDeploymentID: string(experimentDeploymentID),
-	}, nil
-}
-
-// WriteExperiment writes the experiment to the directories.
-func (d *Directories) WriteExperiment(ctx context.Context, operations Operations) error {
-	if runtime.GOOS == "windows" {
-		// On windows, experiments are not supported yet for configuration.
-		return operations.Apply(d.StablePath)
-	}
-	err := os.RemoveAll(d.ExperimentPath)
-	if err != nil {
-		return err
-	}
-	err = copyDirectory(ctx, d.StablePath, d.ExperimentPath)
-	if err != nil {
-		return err
-	}
-
-	operations.FileOperations = append(buildOperationsFromLegacyInstaller(d.StablePath), operations.FileOperations...)
-
-	err = operations.Apply(d.ExperimentPath)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// PromoteExperiment promotes the experiment to the stable.
-func (d *Directories) PromoteExperiment(_ context.Context) error {
-	if runtime.GOOS == "windows" {
-		// On windows, experiments are not supported yet for configuration.
-		return nil
-	}
-	// check if experiment path exists using os
-	_, err := os.Stat(d.ExperimentPath)
-	if err != nil {
-		return err
-	}
-	err = replaceConfigDirectory(d.StablePath, d.ExperimentPath)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// RemoveExperiment removes the experiment from the directories.
-func (d *Directories) RemoveExperiment(_ context.Context) error {
-	if runtime.GOOS == "windows" {
-		// On windows, experiments are not supported yet for configuration.
-		return nil
-	}
-	err := os.RemoveAll(d.ExperimentPath)
-	if err != nil {
-		return err
-	}
-	err = symlink.Set(d.ExperimentPath, d.StablePath)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 // Operations is the list of operations to perform on the configs.
 type Operations struct {
 	DeploymentID   string          `json:"deployment_id"`
@@ -155,10 +64,6 @@ func (o *Operations) Apply(rootPath string) error {
 		if err != nil {
 			return err
 		}
-	}
-	err = os.WriteFile(filepath.Join(rootPath, deploymentIDFile), []byte(o.DeploymentID), 0640)
-	if err != nil {
-		return err
 	}
 	return nil
 }
