--- pkg/fleet/installer/config/file_nix.go
@@ -1,21 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build !windows
-
-package config
-
-import (
-	"context"
-
-	"github.com/DataDog/datadog-agent/pkg/fleet/installer/telemetry"
-)
-
-// copyDirectory copies a directory from source to target.
-// It preserves the directory structure and file permissions.
-func copyDirectory(ctx context.Context, sourcePath, targetPath string) error {
-	cmd := telemetry.CommandContext(ctx, "cp", "-a", sourcePath, targetPath)
-	return cmd.Run()
-}
