+++ pkg/fleet/installer/paths/installer_paths.go
@@ -21,6 +21,8 @@ const (
 	DefaultUserConfigsDir = "/etc"
 	// AgentConfigDir is the path to the agent configuration directory.
 	AgentConfigDir = "/etc/datadog-agent"
+	// AgentConfigDirExp is the path to the agent configuration directory for experiments.
+	AgentConfigDirExp = "/etc/datadog-agent-exp"
 	// StableInstallerPath is the path to the stable installer binary.
 	StableInstallerPath = "/opt/datadog-packages/datadog-installer/stable/bin/installer/installer"
 	// ExperimentInstallerPath is the path to the experiment installer binary.
