+++ pkg/fleet/installer/paths/installer_paths_windows.go
@@ -46,6 +46,8 @@ var (
 	ConfigsPath string
 	// AgentConfigDir is the path to the agent configuration directory.
 	AgentConfigDir string
+	// AgentConfigDirExp is the path to the agent configuration directory for experiments.
+	AgentConfigDirExp string
 	// RootTmpDir is the temporary path where the bootstrapper will be extracted to.
 	RootTmpDir string
 	// DefaultUserConfigsDir is the default Agent configuration directory
@@ -85,6 +87,7 @@ func init() {
 	}
 	AgentConfigDir = DatadogDataDir
 	DatadogInstallerData = filepath.Join(DatadogDataDir, "Installer")
+	AgentConfigDirExp = DatadogDataDir + "-exp"
 	PackagesPath = filepath.Join(DatadogInstallerData, "packages")
 	ConfigsPath = filepath.Join(DatadogInstallerData, "managed")
 	RootTmpDir = filepath.Join(DatadogInstallerData, "tmp")
